Adopt cargo workspace
