+++ Cargo.toml
@@ -1,48 +1,11 @@
-[package]
-name = "chalamet_pir"
-version = "0.6.0"
-edition = "2024"
-resolver = "2"
-rust-version = "1.85.0"
-authors = ["Anjan Roy <hello@itzmeanjan.in>"]
-description = "Simple, Stateful, Single-Server Private Information Retrieval for Key-Value Databases"
-readme = "README.md"
-repository = "https://github.com/itzmeanjan/ChalametPIR.git"
-license = "MPL-2.0"
-keywords = [
-    "priv-info-retrieval",
-    "lwe-pir",
-    "frodo-pir",
-    "chalamet-pir",
-    "gpu",
+[workspace]
+members = [
+    "chalamet_pir_client",
+    "chalamet_pir_common",
+    "chalamet_pir_server",
+    "integrations",
 ]
-categories = ["cryptography", "data-structures", "concurrency"]
-
-[dependencies]
-turboshake = "=0.4.1"
-rayon = "=1.10.0"
-rand = "=0.9.1"
-rand_chacha = "=0.9.0"
-vulkano = { version = "=0.35.1", optional = true }
-vulkano-shaders = { version = "=0.35.0", optional = true }
-
-[dev-dependencies]
-test-case = "=3.3.1"
-divan = "=0.1.21"
-unicode-xid = "=0.2.6"
-
-[[bench]]
-name = "offline_phase"
-harness = false
-
-[[bench]]
-name = "online_phase"
-harness = false
-required-features = ["mutate_internal_client_state"]
-
-[features]
-mutate_internal_client_state = []
-gpu = ["dep:vulkano", "dep:vulkano-shaders"]
+resolver = "3"
 
 [profile.optimized]
 inherits = "release"
