+++ chalamet_pir_client/src/lib.rs
@@ -1,0 +1,4 @@
+mod client;
+
+pub use chalamet_pir_common::{error::ChalametPIRError, params::SEED_BYTE_LEN};
+pub use client::{Client, Query};
