+++ chalamet_pir_common/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "chalamet_pir_common"
+version = "0.6.0"
+edition = "2024"
+resolver = "3"
+rust-version = "1.85.0"
+authors = ["Anjan Roy <hello@itzmeanjan.in>"]
+description = "Simple, Stateful, Single-Server Private Information Retrieval for Key-Value Databases"
+readme = "README.md"
+repository = "https://github.com/itzmeanjan/ChalametPIR.git"
+license = "MPL-2.0"
+keywords = [
+    "priv-info-retrieval",
+    "lwe-pir",
+    "frodo-pir",
+    "chalamet-pir",
+    "gpu",
+]
+categories = ["cryptography", "data-structures", "concurrency"]
+
+[dependencies]
+turboshake = "=0.4.1"
+rand = "=0.9.1"
+rand_chacha = "=0.9.0"
+rayon = "=1.10.0"
+
+[dev-dependencies]
+test-case = "=3.3.1"
