+++ chalamet_pir_common/src/branch_opt_util.rs
