+++ chalamet_pir_common/src/error.rs
