+++ chalamet_pir_common/src/matrix.rs
@@ -1,5 +1,6 @@
-use crate::pir_internals::{
+use crate::{
     binary_fuse_filter, branch_opt_util,
+    error::ChalametPIRError,
     params::{HASHED_KEY_BYTE_LEN, MAX_CIPHER_TEXT_BIT_LEN, MIN_CIPHER_TEXT_BIT_LEN, SEED_BYTE_LEN},
     serialization,
 };
@@ -14,8 +15,6 @@ use turboshake::TurboShake128;
 
 #[cfg(test)]
 use std::ops::Neg;
-
-use super::error::ChalametPIRError;
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct Matrix {
@@ -90,7 +89,7 @@ impl Matrix {
     }
 
     pub fn row_wise_compress(self, mat_elem_bit_len: usize) -> Result<Matrix, ChalametPIRError> {
-        if branch_opt_util::unlikely(!(mat_elem_bit_len >= MIN_CIPHER_TEXT_BIT_LEN && mat_elem_bit_len <= MAX_CIPHER_TEXT_BIT_LEN)) {
+        if branch_opt_util::unlikely(!(MIN_CIPHER_TEXT_BIT_LEN..=MAX_CIPHER_TEXT_BIT_LEN).contains(&mat_elem_bit_len)) {
             return Err(ChalametPIRError::ImpossibleEncodedDBMatrixElementBitLength);
         }
 
@@ -1092,55 +1091,19 @@ impl Neg for &Matrix {
 }
 
 #[cfg(test)]
-pub mod test {
+mod test {
     use crate::{
-        SEED_BYTE_LEN,
-        pir_internals::{
-            binary_fuse_filter::BinaryFuseFilter,
-            error::ChalametPIRError,
-            matrix::Matrix,
-            params::{MAX_CIPHER_TEXT_BIT_LEN, MIN_CIPHER_TEXT_BIT_LEN, SERVER_SETUP_MAX_ATTEMPT_COUNT},
-        },
+        binary_fuse_filter::BinaryFuseFilter,
+        error::ChalametPIRError,
+        matrix::Matrix,
+        params::SEED_BYTE_LEN,
+        params::{MAX_CIPHER_TEXT_BIT_LEN, MIN_CIPHER_TEXT_BIT_LEN, SERVER_SETUP_MAX_ATTEMPT_COUNT},
+        utils::generate_random_kv_database,
     };
     use rand::prelude::*;
     use rand_chacha::ChaCha8Rng;
     use std::collections::HashMap;
     use test_case::test_case;
-
-    /// Generates a random key-value database with the requested number of key-value pairs.
-    ///
-    /// # Arguments
-    ///
-    /// * `num_kv_pairs` - The number of key-value pairs to generate.
-    ///
-    /// # Returns
-    ///
-    /// * `HashMap<Vec<u8>, Vec<u8>>` - A HashMap containing the generated key-value pairs.
-    ///   The keys and values are randomly generated byte arrays with lengths between fixed minimum and maximum values.
-    pub fn generate_random_kv_database(num_kv_pairs: usize) -> HashMap<Vec<u8>, Vec<u8>> {
-        const MIN_KEY_BYTE_LEN: usize = 16;
-        const MAX_KEY_BYTE_LEN: usize = 32;
-        const MIN_VALUE_BYTE_LEN: usize = 1;
-        const MAX_VALUE_BYTE_LEN: usize = 512;
-
-        let mut kv = HashMap::with_capacity(num_kv_pairs);
-        let mut rng = ChaCha8Rng::from_os_rng();
-
-        for _ in 0..num_kv_pairs {
-            let key_byte_len = rng.random_range(MIN_KEY_BYTE_LEN..=MAX_KEY_BYTE_LEN);
-            let value_byte_len = rng.random_range(MIN_VALUE_BYTE_LEN..=MAX_VALUE_BYTE_LEN);
-
-            let mut key = vec![0u8; key_byte_len];
-            let mut value = vec![0u8; value_byte_len];
-
-            rng.fill_bytes(&mut key);
-            rng.fill_bytes(&mut value);
-
-            kv.insert(key, value);
-        }
-
-        kv
-    }
 
     #[test]
     fn encode_kv_database_using_3_wise_xor_filter_and_recover_values() {
