+++ chalamet_pir_server/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "chalamet_pir_server"
+version = "0.6.0"
+edition = "2024"
+resolver = "3"
+rust-version = "1.85.0"
+authors = ["Anjan Roy <hello@itzmeanjan.in>"]
+description = "Simple, Stateful, Single-Server Private Information Retrieval for Key-Value Databases"
+readme = "README.md"
+repository = "https://github.com/itzmeanjan/ChalametPIR.git"
+license = "MPL-2.0"
+keywords = [
+    "priv-info-retrieval",
+    "lwe-pir",
+    "frodo-pir",
+    "chalamet-pir",
+    "gpu",
+]
+categories = ["cryptography", "data-structures", "concurrency"]
+
+[dependencies]
+chalamet_pir_common = { path = "../chalamet_pir_common" }
+vulkano = { version = "=0.35.1", optional = true }
+vulkano-shaders = { version = "=0.35.0", optional = true }
+
+[features]
+gpu = ["dep:vulkano", "dep:vulkano-shaders"]
