+++ integrations/benches/offline_phase.rs
@@ -1,8 +1,11 @@
-use chalamet_pir::{client, server};
+use std::{collections::HashMap, time::Duration};
+
+use chalamet_pir_client::Client;
+use chalamet_pir_server::Server;
+
 use divan;
 use rand::prelude::*;
 use rand_chacha::ChaCha8Rng;
-use std::{collections::HashMap, time::Duration};
 
 fn main() {
     divan::main();
@@ -60,12 +63,12 @@ fn server_setup<const ARITY: u32>(bencher: divan::Bencher, db_config: &DBConfig)
     let kv = generate_random_kv_database(&mut rng, db_config.db_entry_count, db_config.key_byte_len, db_config.value_byte_len);
     let kv_as_ref = kv.iter().map(|(k, v)| (k.as_slice(), v.as_slice())).collect::<HashMap<&[u8], &[u8]>>();
 
-    let mut seed_μ = [0u8; chalamet_pir::SEED_BYTE_LEN];
+    let mut seed_μ = [0u8; chalamet_pir_server::SEED_BYTE_LEN];
     rng.fill_bytes(&mut seed_μ);
 
     bencher
         .with_inputs(|| (kv_as_ref.clone(), seed_μ))
-        .bench_values(|(kv, seed)| server::Server::setup::<ARITY>(divan::black_box(&seed), divan::black_box(kv)));
+        .bench_values(|(kv, seed)| Server::setup::<ARITY>(divan::black_box(&seed), divan::black_box(kv)));
 }
 
 #[divan::bench(args = ARGS, consts = ARITIES, max_time = Duration::from_secs(300), skip_ext_time = true)]
@@ -75,9 +78,9 @@ fn client_setup<const ARITY: u32>(bencher: divan::Bencher, db_config: &DBConfig)
     let kv = generate_random_kv_database(&mut rng, db_config.db_entry_count, db_config.key_byte_len, db_config.value_byte_len);
     let kv_as_ref = kv.iter().map(|(k, v)| (k.as_slice(), v.as_slice())).collect::<HashMap<&[u8], &[u8]>>();
 
-    let mut seed_μ = [0u8; chalamet_pir::SEED_BYTE_LEN];
+    let mut seed_μ = [0u8; chalamet_pir_server::SEED_BYTE_LEN];
     rng.fill_bytes(&mut seed_μ);
 
-    let (_, hint_bytes, filter_param_bytes) = server::Server::setup::<ARITY>(&seed_μ, kv_as_ref).expect("Server setup failed");
-    bencher.bench(|| client::Client::setup(divan::black_box(&seed_μ), divan::black_box(&hint_bytes), divan::black_box(&filter_param_bytes)));
+    let (_, hint_bytes, filter_param_bytes) = Server::setup::<ARITY>(&seed_μ, kv_as_ref).expect("Server setup failed");
+    bencher.bench(|| Client::setup(divan::black_box(&seed_μ), divan::black_box(&hint_bytes), divan::black_box(&filter_param_bytes)));
 }
