+++ integrations/benches/online_phase.rs
@@ -1,8 +1,11 @@
-use chalamet_pir::{client, server};
+use std::{collections::HashMap, time::Duration};
+
+use chalamet_pir_client::Client;
+use chalamet_pir_server::Server;
+
 use divan;
 use rand::prelude::*;
 use rand_chacha::ChaCha8Rng;
-use std::{collections::HashMap, time::Duration};
 
 fn main() {
     divan::main();
@@ -60,11 +63,11 @@ fn client_query<const ARITY: u32>(bencher: divan::Bencher, db_config: &DBConfig)
     let kv = generate_random_kv_database(&mut rng, db_config.db_entry_count, db_config.key_byte_len, db_config.value_byte_len);
     let kv_as_ref = kv.iter().map(|(k, v)| (k.as_slice(), v.as_slice())).collect::<HashMap<&[u8], &[u8]>>();
 
-    let mut seed_μ = [0u8; chalamet_pir::SEED_BYTE_LEN];
+    let mut seed_μ = [0u8; chalamet_pir_server::SEED_BYTE_LEN];
     rng.fill_bytes(&mut seed_μ);
 
-    let (_, hint_bytes, filter_param_bytes) = server::Server::setup::<ARITY>(&seed_μ, kv_as_ref.clone()).unwrap();
-    let client = client::Client::setup(&seed_μ, &hint_bytes, &filter_param_bytes).unwrap();
+    let (_, hint_bytes, filter_param_bytes) = Server::setup::<ARITY>(&seed_μ, kv_as_ref.clone()).unwrap();
+    let client = Client::setup(&seed_μ, &hint_bytes, &filter_param_bytes).unwrap();
 
     let (&key, _) = kv_as_ref.iter().last().unwrap();
 
@@ -81,11 +84,11 @@ fn server_respond<const ARITY: u32>(bencher: divan::Bencher, db_config: &DBConfi
     let kv = generate_random_kv_database(&mut rng, db_config.db_entry_count, db_config.key_byte_len, db_config.value_byte_len);
     let kv_as_ref = kv.iter().map(|(k, v)| (k.as_slice(), v.as_slice())).collect::<HashMap<&[u8], &[u8]>>();
 
-    let mut seed_μ = [0u8; chalamet_pir::SEED_BYTE_LEN];
+    let mut seed_μ = [0u8; chalamet_pir_server::SEED_BYTE_LEN];
     rng.fill_bytes(&mut seed_μ);
 
-    let (server, hint_bytes, filter_param_bytes) = server::Server::setup::<ARITY>(&seed_μ, kv_as_ref.clone()).unwrap();
-    let mut client = client::Client::setup(&seed_μ, &hint_bytes, &filter_param_bytes).unwrap();
+    let (server, hint_bytes, filter_param_bytes) = Server::setup::<ARITY>(&seed_μ, kv_as_ref.clone()).unwrap();
+    let mut client = Client::setup(&seed_μ, &hint_bytes, &filter_param_bytes).unwrap();
 
     let (&key, _) = kv_as_ref.iter().last().unwrap();
     let query_bytes = client.query(key).unwrap();
@@ -100,11 +103,11 @@ fn client_process_response<const ARITY: u32>(bencher: divan::Bencher, db_config:
     let kv = generate_random_kv_database(&mut rng, db_config.db_entry_count, db_config.key_byte_len, db_config.value_byte_len);
     let kv_as_ref = kv.iter().map(|(k, v)| (k.as_slice(), v.as_slice())).collect::<HashMap<&[u8], &[u8]>>();
 
-    let mut seed_μ = [0u8; chalamet_pir::SEED_BYTE_LEN];
+    let mut seed_μ = [0u8; chalamet_pir_server::SEED_BYTE_LEN];
     rng.fill_bytes(&mut seed_μ);
 
-    let (server, hint_bytes, filter_param_bytes) = server::Server::setup::<ARITY>(&seed_μ, kv_as_ref.clone()).unwrap();
-    let mut client = client::Client::setup(&seed_μ, &hint_bytes, &filter_param_bytes).unwrap();
+    let (server, hint_bytes, filter_param_bytes) = Server::setup::<ARITY>(&seed_μ, kv_as_ref.clone()).unwrap();
+    let mut client = Client::setup(&seed_μ, &hint_bytes, &filter_param_bytes).unwrap();
 
     let (&key, _) = kv_as_ref.iter().last().unwrap();
     let query_bytes = client.query(key).unwrap();
