+++ integrations/src/lib.rs
@@ -1,0 +1,1 @@
+mod test_pir;
